+++ src/const.js
@@ -1,7 +1,7 @@
 const KEY_CODE_ESCAPE = 27;
 
 const FILMS_QUANTITY = 25;
-const FILMS_COUNT_PER_STEP = 5;
+const CARDS_COUNT_PER_STEP = 5;
 const SPECIAL_CARDS_COUNT = 2;
 const MAX_DESCRIPTION_LENGTH = 140;
 
@@ -9,11 +9,6 @@ const CLICKABLE_HTML_ELEMENTS = {
   IMG: `IMG`,
   H3: `H3`,
   A: `A`
-};
-
-const SORTING_OPTION_CLASSES = {
-  DEFAULT: `sort__button`,
-  ACTIVE: `sort__button--active`,
 };
 
 const SORTING_ENTRIES = {
@@ -60,11 +55,10 @@ const PROFILE_RATING = {
 export {
   KEY_CODE_ESCAPE,
   FILMS_QUANTITY,
-  FILMS_COUNT_PER_STEP,
+  CARDS_COUNT_PER_STEP,
   SPECIAL_CARDS_COUNT,
   MAX_DESCRIPTION_LENGTH,
   CLICKABLE_HTML_ELEMENTS,
-  SORTING_OPTION_CLASSES,
   FILTER_OPTION_CLASSES,
   FILTER_ENTRIES,
   SORTING_ENTRIES,
