+++ src/utils/common.js
@@ -33,16 +33,4 @@ export const getNumberFromString = (string) => {
   return parseInt(number, 10);
 };
 
-export const updateItem = (items, updatedItem) => {
-  const index = items.findIndex((item) => item.id === updatedItem.id);
-
-  if (index === -1) {
-    return items;
-  }
-
-  return [
-    ...items.slice(0, index),
-    updatedItem,
-    ...items.slice(index + 1)
-  ];
-};
+export const generateId = () => Date.now() + parseInt(Math.random() * 10000, 10);
